rename command modules
