+++ duzzy-editor/src/command/switch.rs
