+++ duzzy-editor/src/lib.rs
@@ -8,8 +8,8 @@ mod buffer;
 mod command;
 mod document;
 pub mod editor;
+mod event;
 mod history;
-mod input;
 mod keymap;
 pub mod renderer;
 mod selection;
